consumer/requirements.txt
kafka-python==2.0.1
python-dateutil==2.8.1
text-unidecode==1.3
mysql-connector-python